bad turret is bad

+++ application/src/Entity.java
@@ -13,7 +13,7 @@ class Entity
 	//protected Point position;
 	protected int posX;
 	protected int posY;
-	protected int direction;
+	protected float direction;
 	protected Rectangle2D bounds;
